requirements.txt
altair==4.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.2.0
colorama==0.4.4
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
gunicorn==20.1.0
idna==3.3
importlib-metadata==5.1.0
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.56.4
numpy==1.23.5
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.26.0
rich==12.6.0
scikit-learn==1.0.2
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
streamlit==1.15.2
streamlit-shap==1.0.2
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.7
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.11.0
